CQA/requirements_windows.txt
absl-py==2.1
addict==2.4
annotated-types==0.7
anyio==4.6
argon2-cffi==23.1
argon2-cffi-bindings==21.2
arrow==1.3
asttokens==3.0
async-lru==2.0
attrs==24.2
babel==2.16
beautifulsoup4==4.12
bleach==6.2
certifi
cffi==1.17
charset-normalizer==3.4
click==8.1
cloudpickle==3.1
colorama==0.4
comm==0.2
contourpy==1.3
cycler==0.12
debugpy==1.8
decorator==5.1
docker-pycreds==0.4
einops==0.8
executing==2.1
fastjsonschema==2.21
filelock==3.16
fire==0.7
fonttools==4.55
fqdn==1.5
fsspec==2024.10
ftfy==6.3
gitdb
GitPython==3.1
-e git+https://github.com/IDEA-Research/GroundingDINO@856dde20aee659246248e20734ef9ba5214f5e44#egg=groundingdino
grpcio==1.68
h11==0.14
h5py==3.12
httpcore==1.0
httpx==0.27
huggingface-hub==0.26
idna==3.10
imageio==2.37
ipykernel==6.29
ipython==8.30
isoduration==20.11
jedi==0.19
Jinja2==3.1
joblib==1.4
json5==0.10
jsonpointer==3.0
jsonschema==4.23
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab
lazy_loader==0.4
loguru==0.7
Markdown==3.7
MarkupSafe==3.0
matplotlib==3.9
matplotlib-inline==0.1
medmnist==3.0
mistune==3.0
mpmath==1.3
nest-asyncio==1.6
networkx==3.4
notebook==7.3
notebook_shim==0.2
numpy>=1.23
ollama==0.4
opencv-python
overrides==7.7
packaging==24.2
pandas
pandocfilters==1.5
parso>=0.8.3
pillow==11.0
platformdirs==4.3
prometheus_client==0.21
prompt_toolkit>=3.0.41
protobuf==5.29
psutil==6.1
pure_eval==0.2
pycparser==2.22
pydantic==2.10
pydantic_core==2.27
Pygments==2.18
pyparsing==3.2
python-dateutil==2.9
python-json-logger>=2.0.4
pytorchcv
pytz==2024.2
pywin32==308
pyzmq==26.2
referencing==0.35
regex
requests==2.32
rfc3339-validator>=0.1.1
rfc3986-validator>=0.1.1
rpds-py==0.22
scikit-image==0.25
scikit-learn==1.5
scipy==1.14
seaborn==0.13
sentry-sdk==2.19
setproctitle==1.3
setuptools==75.1
six==1.17
smmap==5.0
sniffio==1.3
soupsieve==2.6
stack-data==0.6
submitit==1.5
sympy==1.13
termcolor==2.5
terminado==0.18
threadpoolctl==3.5
tifffile
timm
tinycss2==1.4
tokenizers==0.21
tornado==6.4
tqdm==4.67
traitlets==5.14
transformers==4.47
types-python-dateutil
tzdata==2024.2
uri-template==1.3
urllib3==2.2
wandb==0.19
wcwidth==0.2
webcolors==24.11
webencodings==0.5
websocket-client==1.8
Werkzeug==3.1
wheel==0.44
win32-setctime==1.1
yapf==0.43
